+++ crates/slipstore/src/tests.rs
@@ -1,9 +1,0 @@
-use super::*;
-
-#[tokio::test]
-async fn no_panic() {
-    // let db = Database::connect("sqlite::memory:").await.unwrap();
-    // let schema = Schema::new(DbBackend::Sqlite);
-    // let stmt = schema.create_table_from_entity(Foo);
-    // let result = db.execute(db.get_database_backend().build(&stmt)).await;
-}