+++ crates/slipstream/src/feeds/mod.rs
@@ -1,0 +1,13 @@
+//! Feed module.
+
+use super::*;
+
+mod cache;
+mod feed_options;
+mod feeds;
+mod filters;
+
+pub use cache::*;
+pub use feed_options::*;
+pub use feeds::*;
+pub use filters::*;