+++ CHANGELOG.md
@@ -15,3 +15,26 @@
 ### Fixed
 
 ### Security
+
+## 0.3.0
+
+### Added
+
+- Added web views to `slipstream`.
+- Added tag filters to `slipstream`.
+
+### Changed
+
+- Merged `slipknot` into `slipstream`.
+
+### Deprecated
+
+### Removed
+
+- `slipknot` has been removed.
+
+### Fixed
+
+- The `[all]` config section is now respected.
+
+### Security